refactor: update existing code to use new config implementation
