+++ src/config/env.schema.ts
@@ -1,0 +1,14 @@
+import * as Joi from 'joi';
+
+export const EnvSchema = Joi.object({
+  NODE_ENV: Joi.string()
+    .valid('development', 'production', 'test')
+    .default('development'),
+  PORT: Joi.number().port().default(3000),
+  DB_HOST: Joi.string().default('localhost'),
+  DB_PORT: Joi.number().port().default(5432),
+  DB_USERNAME: Joi.string().default('postgres'),
+  DB_PASSWORD: Joi.string().default('password'),
+  DB_NAME: Joi.string().default('nest'),
+  DB_SSL: Joi.boolean().default(false),
+});
