--- src/config/envs/development.ts
@@ -1,8 +1,0 @@
-export const config = {
-  app: {
-    environment: 'development',
-  },
-  db: {
-    logging: true,
-  },
-};
