--- src/config/envs/production.ts
@@ -1,5 +1,0 @@
-export const config = {
-  app: {
-    environment: 'production',
-  },
-};
